+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 <!-- markdownlint-disable no-trailing-punctuation -->
 
+## next
+
+TODO: Date
+
+- Potentially breaking features:
+  - Added `.into_result()` extension method, which losslessly converts from `Fine` to `Result`.
+
 ## 0.0.1
 
 2021-09-30